Txt2img completely done
This was tested on Windows with CUDAExecutionProvider, DmlExecutionProvider
and CPUExecutionProvider.

Cuda and DirectML provide (at least currently) about the same performance
on my RTX 3080, however this may just be due to my subpar implementation.

CPUExecutionProvider is faster on native Windows than on WSL. (by about 25%)
In practice, on my Ryzen 9 5900X, WSL is 20s/it and Windows is 15s/it
(may have been a fluke, who knows at this point)